[BE] Fix: 몽깅이 강화된 스탯 조회 쿼리 수정 (S13P21D103-167)
* fix: 몽깅이 강화된 스탯 조회 쿼리 수정

+++ api-server/src/main/java/com/ggumtle/ggumtle/messaging/payload/RequestRoomPayload.java
@@ -16,9 +16,9 @@ public record RequestRoomPayload(
                     po.memberId(),
                     po.nickname(),
                     po.monggingClassId(),
-                    po.monggingClassId() == 1 ? po.additionalStat() : 0,
-                    po.monggingClassId() == 2 ? po.additionalStat() : 0,
-                    po.monggingClassId() == 3 ? po.additionalStat() : 0
+                    (int) (po.monggingClassId() == 1 ? po.additionalStat() : 0),
+                    (int) (po.monggingClassId() == 2 ? po.additionalStat() : 0),
+                    (int) (po.monggingClassId() == 3 ? po.additionalStat() : 0)
             );
 
             players.add(player);
