+++ api-server/src/main/java/com/ggumtle/ggumtle/mongging/persistence/po/MonggingStatPo.java
@@ -6,6 +6,6 @@ public record MonggingStatPo(
         Long monggingId,
         Integer monggingLevel,
         Long monggingClassId,
-        Integer additionalStat
+        Double additionalStat
 ) {
 }
